Move sample context under the `context` package
