+++ server/src/test/kotlin/io/spine/server/route/EventReactionRoutingSpec.kt
@@ -25,15 +25,14 @@
  */
 package io.spine.server.route
 
-import io.spine.server.given.users.RSessionId
-import io.spine.server.given.users.SessionProjection
-import io.spine.server.given.users.createUsersContext
-import io.spine.server.given.users.event.rUserSignedIn
-import io.spine.server.given.users.rSession
+import io.spine.server.given.context.users.RSessionId
+import io.spine.server.given.context.users.SessionProjection
+import io.spine.server.given.context.users.createUsersContext
+import io.spine.server.given.context.users.event.rUserSignedIn
+import io.spine.server.given.context.users.rSession
 import io.spine.server.testing.blackbox.assertEntity
 import io.spine.testing.core.given.GivenUserId
 import io.spine.testing.server.blackbox.BlackBox
-import org.junit.jupiter.api.Disabled
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.RepeatedTest
 
@@ -53,7 +52,7 @@ internal class EventReactionRoutingSpec {
      * out `@Disabled` annotation for details.
      */
     @RepeatedTest(50) //@Test
-    @Disabled // See https://github.com/SpineEventEngine/core-java/issues/925.
+    //@Disabled // See https://github.com/SpineEventEngine/core-java/issues/925.
     @DisplayName("only occur after the origin event has been already dispatched")
     fun occurAfterOriginDispatched() {
         val userId = GivenUserId.generated()
